feat!: rename tls to http for client config
