+++ htsget-config/src/config/mod.rs
@@ -10,8 +10,8 @@ use crate::config::service_info::ServiceInfo;
 use crate::config::ticket_server::TicketServerConfig;
 use crate::error::Error::{ArgParseError, ParseError, TracingError};
 use crate::error::Result;
+use crate::http::KeyPairScheme;
 use crate::storage::Backend;
-use crate::tls::KeyPairScheme;
 use clap::{Args as ClapArgs, Command, FromArgMatches, Parser};
 use http::header::AUTHORIZATION;
 use http::uri::Authority;
@@ -368,8 +368,8 @@ pub(crate) mod tests {
   use crate::config::advanced::auth::AuthMode;
   use crate::config::location::Location;
   use crate::config::parser::from_str;
+  use crate::http::tests::with_test_certificates;
   use crate::storage::Backend;
-  use crate::tls::tests::with_test_certificates;
   use crate::types::Scheme;
   use figment::Jail;
   use http::Uri;
