+++ htsget-config/src/lib.rs
@@ -4,7 +4,7 @@ pub mod config;
 #[cfg(feature = "experimental")]
 pub mod encryption_scheme;
 pub mod error;
+pub mod http;
 pub mod resolver;
 pub mod storage;
-pub mod tls;
 pub mod types;
